api/requirements.txt
flask
flask-cors
mysql-connector-python
Flask-JWT-Extended